requirements.txt
pytesseract==0.3.8
pillow
